06-lab-complete/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pyjwt==2.9.0
python-dotenv==1.0.1
redis==5.1.0
psutil==6.0.0
# LangGraph + LangChain (Lab 4 TravelBuddy Agent)
langchain-core>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
